# 🚀 Epic 3: Complete UX Dashboard Improvements (9/9 tasks)