--- a/app/assets/stylesheets/application.css
+++ b/app/assets/stylesheets/application.css
@@ -11,4 +11,5 @@
 
 /*
  *= require tailwind/application
+ *= require components/view_toggle
  */