--- a/app/views/expenses/_inline_actions.html.erb
+++ b/app/views/expenses/_inline_actions.html.erb
@@ -1,0 +1,2 @@
+<%= link_to "Ver", expense, class: "text-teal-600 hover:text-teal-800" %>
+<%= link_to "Editar", edit_expense_path(expense), class: "text-slate-600 hover:text-slate-800" %>