… .claude/settings.local.json
… BUG_REPORT_TASK_3_1.md
… CLAUDE.md
… app/javascript/controllers/accessibility_enhanced_controller.js
… app/javascript/controllers/batch_selection_controller.js
… app/javascript/controllers/bulk_operations_controller.js
… app/javascript/controllers/filter_persistence_controller.js
… app/javascript/controllers/inline_actions_controller.js
… app/javascript/controllers/simple_inline_actions_controller.js
… app/javascript/controllers/toast_container_controller.js
… app/javascript/controllers/virtual_scroll_controller.js
… app/jobs/bulk_deletion_job.rb
… app/jobs/bulk_operations/base_job.rb
… app/jobs/bulk_status_update_job.rb
… app/middleware/performance_monitoring_middleware.rb
… app/models/concerns/expense_query_optimizer.rb
… app/models/concerns/query_security.rb
… app/services/bulk_operations/base_service.rb
… app/services/bulk_operations/categorization_service.rb
… app/services/bulk_operations/deletion_service.rb
… app/services/bulk_operations/status_update_service.rb
… app/services/categorization/bulk_categorization_service.rb
… app/services/expense_filter_service.rb
… app/views/expenses/_bulk_operations_modal.html.erb
… app/views/expenses/_expense_row.html.erb
… app/views/expenses/_status_badge.html.erb
… app/views/expenses/index.html.erb
… app/views/layouts/application.html.erb
… config/database.yml
… config/initializers/categorization.rb